Promote SystemDrawing adapter types
