+++ src/WinPrint.Core/Abstractions/SystemDrawingPen.cs
@@ -1,0 +1,27 @@
+using System;
+using System.Drawing;
+using System.Drawing.Drawing2D;
+using System.Drawing.Text;
+
+namespace WinPrint.Core.Abstractions;
+
+internal sealed class SystemDrawingPen : IGraphicsPen
+{
+    private readonly bool _ownsNative;
+
+    public SystemDrawingPen (Pen pen, bool ownsNative = true)
+    {
+        Pen = pen ?? throw new ArgumentNullException (nameof (pen));
+        _ownsNative = ownsNative;
+    }
+
+    public Pen Pen { get; }
+
+    public void Dispose ()
+    {
+        if (_ownsNative)
+        {
+            Pen.Dispose ();
+        }
+    }
+}
